+++ bloc_generator/example/example.dart
@@ -1,1 +1,0 @@
-/// An extensive example is available at https://github.com/CallumIddon/flutter_bloc_generator/example