+++ bloc_generator/lib/src/metadata.dart
@@ -1,4 +1,18 @@
-import "package:analyzer/dart/element/element.dart";
+// Copyright 2019 Callum Iddon
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import 'package:analyzer/dart/element/element.dart';
 
 bool findMetadata(Element element, String metadataName) {
   bool foundMetadata = false;
@@ -13,9 +27,9 @@ bool findMetadata(Element element, String metadataName) {
 
 List<ElementAnnotation> getMetadata(Element element, String metadataName) {
   if (!findMetadata(element, metadataName)) {
-    throw Exception("$element does not have a $metadataName");
+    throw Exception('$element does not have a $metadataName');
   }
-  List<ElementAnnotation> foundMetadata = <ElementAnnotation>[];
+  final List<ElementAnnotation> foundMetadata = <ElementAnnotation>[];
   for (ElementAnnotation metadata in element.metadata) {
     if (metadata.toSource().startsWith(metadataName)) {
       foundMetadata.add(metadata);
@@ -26,43 +40,42 @@ List<ElementAnnotation> getMetadata(Element element, String metadataName) {
 
 List<String> findInputs(ElementAnnotation metadata) => metadata
         .toSource()
-        .split("(")[1]
-        .split(")")[0]
-        .split(", ")
+        .split('(')[1]
+        .split(')')[0]
+        .split(', ')
         .map((String input) {
       String output = input;
-      if (output.startsWith("\"")) {
+      if (output.startsWith('\'')) {
         output = output.substring(1);
       }
-      if (output.endsWith("\"")) {
+      if (output.endsWith('\'')) {
         output = output.substring(0, output.length - 1);
       }
       return output;
     }).toList();
 
-bool checkMethod(Element element) => element.toString().indexOf("(") != -1;
+bool checkMethod(Element element) => element.toString().contains('(');
 
 String findName(Element element) {
-  String name = element.toString();
+  final String name = element.toString();
   if (checkMethod(element)) {
-    return name.substring(0, name.indexOf("("));
+    return name.substring(0, name.indexOf('('));
   }
-  return name.split(" ").removeLast();
+  return name.split(' ').removeLast();
 }
 
 String findType(Element element) {
   if (checkMethod(element)) {
-    String type = element.toString().split("→").reversed.toList()[0];
+    final String type = element.toString().split('→').reversed.toList()[0];
     return type.substring(1);
   }
-  final List<String> name = element.toString().split(" ").reversed.toList();
-  name.removeAt(0);
-  return name.join(" ");
+  final List<String> name = element.toString().split(' ').reversed.toList()
+    ..removeAt(0);
+  return name.join(' ');
 }
 
 String findTemplateType(Element element) {
-  List<String> typeList = findType(element).split("<");
-  typeList.removeAt(0);
-  String type = typeList.join("<");
+  final List<String> typeList = findType(element).split('<')..removeAt(0);
+  final String type = typeList.join('<');
   return type.substring(0, type.length - 1);
 }