+++ flutter_bloc_provider/CHANGELOG.md
@@ -1,3 +1,12 @@
+# Changelog
+
+## 0.5.0
+
+* Add Documentation
+* Refactor
+* Format
+* Add CI
+
 ## 0.4.0
 
 * flutter_bloc packages moved to bloc